Removing aggregate from preprocessing

+++ divina/pipeline/model.py
@@ -76,7 +76,7 @@ class GLM(BaseEstimator):
 
     def fit(self, X, y, drop_constants: bool = False):
         if drop_constants:
-            da_std = X.std()
+            da_std = X.astype(float).std()
             constant_indices = [i for i, v in enumerate(da_std) if v == 0]
             usable_indices = list(set(range(len(X.columns))) - set(constant_indices))
             if usable_indices == []:
